+++ keyboards/momokai/tap_trio_pro/lut.h
@@ -3,10 +3,10 @@ SPDX-License-Identifier: GPL-2.0-or-later */
 #pragma once
 #include <stdint.h>
 
-extern uint8_t lut[4096];
+extern uint16_t lut[4096];
 
-uint16_t distance_to_adc(uint8_t distance);
+uint16_t distance_to_adc(uint16_t distance);
 
-uint8_t adc_to_distance(uint16_t adc);
+uint16_t adc_to_distance(uint16_t adc);
 
 void generate_lut(void);
